test: Statement prints transactions

+++ src/main/java/tech/makers/bank/BankAccount.java
@@ -3,24 +3,24 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 
 public class BankAccount {
-    private float balance;
+    private double balance;
     private ArrayList<String> transactions = new ArrayList();
 
     public BankAccount() {
         this.balance = 0;
     }
 
-    public float getBalance() {
+    public double getBalance() {
         return balance;
     }
 
-    private void checkValid(float amount) {
+    private void checkValid(double amount) {
         if (amount < 0 || amount * 100 % 1 != 0) {
             throw new ArithmeticException("Invalid amount");
         }
     }
 
-    public void deposit(float amount, LocalDate date) {
+    public void deposit(double amount, LocalDate date) {
         checkValid(amount);
 
         balance += amount;
@@ -28,13 +28,13 @@ public class BankAccount {
         transactions.add("deposit");
     }
 
-    private void checkFunds(float amount) {
+    private void checkFunds(double amount) {
         if (amount > balance) {
             throw new ArithmeticException("Not enough funds");
         }
     }
 
-    public void withdraw(float amount, LocalDate date) {
+    public void withdraw(double amount, LocalDate date) {
         checkValid(amount);
         checkFunds(amount);
